refactor: proto定義をDDDのレイヤーから切り離す
