+++ Makefile
@@ -20,10 +20,10 @@ lint:
 
 .PHONY: gen_proto
 gen_proto:
-	protoc -I ./interfaces/proto \
-	--go-grpc_out=./interfaces/proto/pb \
+	protoc -I ./proto \
+	--go-grpc_out=./proto/pb \
 	--go-grpc_opt=require_unimplemented_servers=false \
-	--go-grpc_opt=module=github.com/karamaru-alpha/chat-go-server/interfaces/proto \
-	--go_out=./interfaces/proto/pb \
-	--go_opt=module=github.com/karamaru-alpha/chat-go-server/interfaces/proto \
-	./interfaces/proto/*.proto
+	--go-grpc_opt=module=github.com/karamaru-alpha/chat-go-server/proto \
+	--go_out=./proto/pb \
+	--go_opt=module=github.com/karamaru-alpha/chat-go-server/proto \
+	./proto/*.proto
